+++ src/libs/vtools/visualization/line/operation/vistoolrotation.cpp
@@ -141,11 +141,11 @@ void VisToolRotation::RefreshGeometry()
                 const QSharedPointer<VPointF> p = Visualization::data->GeometricObject<VPointF>(id);
 
                 ++iPoint;
-                QGraphicsEllipseItem *point = GetPoint(iPoint, supportColor2);
+                QGraphicsEllipseItem *point = GetPoint(static_cast<quint32>(iPoint), supportColor2);
                 DrawPoint(point, *p, supportColor2);
 
                 ++iPoint;
-                point = GetPoint(iPoint, supportColor);
+                point = GetPoint(static_cast<quint32>(iPoint), supportColor);
 
                 if (object1Id != NULL_ID)
                 {
@@ -216,11 +216,11 @@ int VisToolRotation::AddCurve(qreal angle, const QPointF &origin, quint32 id, in
     const QSharedPointer<Item> curve = Visualization::data->template GeometricObject<Item>(id);
 
     ++i;
-    QGraphicsPathItem *path = GetCurve(i, supportColor2);
+    QGraphicsPathItem *path = GetCurve(static_cast<quint32>(i), supportColor2);
     DrawPath(path, curve->GetPath(PathDirection::Show), supportColor2, Qt::SolidLine, Qt::RoundCap);
 
     ++i;
-    path = GetCurve(i, supportColor);
+    path = GetCurve(static_cast<quint32>(i), supportColor);
     if (object1Id != NULL_ID)
     {
         const Item rotated = curve->Rotate(origin, angle);
